introduce @calls/common; move over types and rtc_stats for now
