+++ standalone/package.json
@@ -76,5 +76,8 @@
     "semver-parser": "4.0.0",
     "styled-components": "5.3.5",
     "typescript": "4.3.4"
-  }
+  },
+  "workspaces": [
+      "../webapp/packages/common"
+  ]
 }
