+++ webapp/packages/common/package.json
@@ -1,0 +1,13 @@
+{
+    "name": "@calls/common",
+    "version": "0.12.0",
+    "description": "Common code shared between calls webapp, desktop, and mobile.",
+    "license": "MIT",
+    "main": "./lib/index.js",
+    "devDependencies": {},
+    "scripts": {
+        "build": "tsc --build --verbose",
+        "run": "tsc --watch --preserveWatchOutput",
+        "clean": "rm -rf tsconfig.tsbuildinfo ./lib"
+    }
+}
