+++ webapp/src/components/emoji/emoji.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 
 import {getEmojiImageUrl} from 'mattermost-redux/utils/emoji_utils';
 
-import {EmojiData} from 'src/types/types';
+import {EmojiData} from '@calls/common';
 import {Emojis, EmojiIndicesByUnicode} from 'src/emojis/emoji';
 
 interface Props {
