+++ webapp/src/components/expanded_view/call_participant.tsx
@@ -6,7 +6,7 @@ import CompassIcon from 'src/components/icons/compassIcon';
 import MutedIcon from 'src/components/icons/muted_icon';
 import UnmutedIcon from 'src/components/icons/unmuted_icon';
 import Avatar from 'src/components/avatar/avatar';
-import {Reaction} from 'src/types/types';
+import {Reaction} from '@calls/common';
 import {Emoji} from 'src/components/emoji/emoji';
 
 export type Props = {
