+++ webapp/src/selectors.ts
@@ -27,7 +27,7 @@ import {getMyChannelRoles, getMyTeamRoles} from 'mattermost-redux/selectors/enti
 import {getMyChannelMemberships} from 'mattermost-redux/selectors/entities/common';
 
 import {getChannelURL} from 'src/utils';
-import {CallsConfig, CallsUserPreferences, ChannelState, Reaction, UserState} from 'src/types/types';
+import {CallsConfig, CallsUserPreferences, ChannelState, Reaction, UserState} from '@calls/common';
 
 import {pluginId} from './manifest';
